% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.inflmlm.R
\name{as.data.frame.inflmlm}
\alias{as.data.frame.inflmlm}
\title{Convert an inflmlm object to a data frame}
\usage{
\method{as.data.frame}{inflmlm}(x, ..., FUN = det, funnames = TRUE)
}
\arguments{
\item{x}{An \code{inflmlm} object, as returned by \code{mlm.influence}}

\item{...}{ignored}

\item{FUN}{in the case where the subset size, \code{m>1}, the function used on the \code{H, Q, L, R} to calculate
a single statistic. The default is \code{det}. An alternative is \code{tr}, for matrix trace.}

\item{funnames}{logical. Should the \code{FUN} name be prepended to the statistics when creating a data frame?}
}
\value{
A data frame containing the influence statistics
}
\description{
This function is used internally in the package to convert the result of \code{mlm.influence()} to a data frame.
It is not normally called by the user.
}
\examples{
# none

}
