% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrPlot.R
\name{lrPlot}
\alias{lrPlot}
\alias{lrPlot.lm}
\title{Regression LR Influence Plot}
\usage{
lrPlot(model, ...)
}
\arguments{
\item{model}{a linear or generalized-linear model.}

\item{\dots}{arguments to pass to the \code{plot} and \code{points}
functions.}
}
\value{
If points are identified, returns a data frame with the hat values,
        Studentized residuals and Cook's distance of the identified points.  If no
        points are identified, nothing is returned.  This function is primarily used
        for its side-effect of drawing a plot.
}
\description{
This function creates a \dQuote{bubble} plot of functions, R =
log(Studentized residuals^2) by L = log(H/p*(1-H)) of the hat values, with
the areas of the circles representing the observations proportional to
Cook's distances.
}
\details{
This plot, suggested by McCulloch & Meeter (1983) has the attractive
property that contours of equal Cook's distance are diagonal lines with
slope = -1.  Various reference lines are drawn on the plot corresponding to
twice and three times the average hat value, a \dQuote{large} squared
studentized residual and contours of Cook's distance.

The \code{id.method="noteworthy"} setting also requires setting
\code{id.n>0} to have any effect. Using \code{id.method="noteworthy"}, and
\code{id.n>0}, the number of points labeled is the union of the largest
\code{id.n} values on each of L, R, and CookD.
}
\examples{

# artificial example from Lawrence (1995)
x <- c( 0, 0, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11, 18, 18 )
y <- c( 0, 6, 6, 7, 6, 7, 6, 7, 6,  7,  6,  7,  7,  18 )
DF <- data.frame(x,y, row.names=LETTERS[1:length(x)])
DF

with(DF, {
	plot(x,y, pch=16, cex=1.3)
	abline(lm(y~x), col="red", lwd=2)
	NB <- c(1,2,13,14)
	text(x[NB],y[NB], LETTERS[NB], pos=c(4,4,2,2))
	}
)

mod <- lm(y~x, data=DF)
# standard influence plot from car
influencePlot(mod, id.n=4)

# lrPlot version
lrPlot(mod, id.n=4)


library(car)
dmod <- lm(prestige ~ income + education, data = Duncan)
influencePlot(dmod, id.n=3)
lrPlot(dmod, id.n=3)

}
\references{
A. J. Lawrence (1995). Deletion Influence and Masking in
Regression \emph{Journal of the Royal Statistical Society. Series B
(Methodological)} , Vol. \bold{57}, No. 1, pp. 181-189.

McCulloch, C. E. & Meeter, D. (1983). Discussion of "Outliers..." by R. J.
Beckman and R. D. Cook.  \emph{Technometrics}, 25, 152-155.
}
\seealso{
\code{\link{influencePlot.mlm}}
\code{\link[car]{influencePlot}} in the \code{car} package for other methods
}
\author{
Michael Friendly
}
\keyword{regression}
