% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inflmlm.R
\name{print.inflmlm}
\alias{print.inflmlm}
\title{Print an inflmlm object}
\usage{
\method{print}{inflmlm}(x, digits = max(3, getOption("digits") - 4), FUN = det, ...)
}
\arguments{
\item{x}{An \code{inflmlm} object}

\item{digits}{Number of digits to print}

\item{FUN}{Function to combine diagnostics when \code{m>1}, one of \code{det} or \code{tr}}

\item{...}{passed to \code{print()}}
}
\value{
Invisibly returns the object
}
\description{
Print an inflmlm object
}
\examples{
# none

}
