% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Matrix trace}
\usage{
tr(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
returns the sum of the diagonal elements of the matrix
}
\description{
Calculates the trace of a matrix
}
\details{
For square, symmetric matrices, such as covariance matrices, the trace is sometimes
used as a measure of size, e.g., in Pillai's trace criterion for a MLM.
}
\examples{

M <- matrix(sample(1:9), 3,3)
tr(M)

}
\author{
Michael Friendly
}
\keyword{array}
